[TOW-50] 마이페이지 기능 추가
* 마이페이지 기능 추가
* 기존 마이페이지 이름 -> 나의 활동 으로 변경
* 활동 -> 마이페이지로 이동 로직 추가
