+++ TodayWod/Features/Tab/AppFeature.swift
@@ -14,13 +14,13 @@ struct AppFeature {
     @ObservableState
     struct State: Equatable, Sendable {
         var homeTab = HomeFeature.State()
-        var settingsTab = MyPageFeature.State()
+        var settingsTab = MyActivityFeature.State()
         var selectedItem: TabMenuItem = .home
     }
 
     enum Action: BindableAction {
         case homeTab(HomeFeature.Action)
-        case settingsTab(MyPageFeature.Action)
+        case settingsTab(MyActivityFeature.Action)
         case resetOnboarding
         case binding(BindingAction<State>)
 
@@ -32,7 +32,7 @@ struct AppFeature {
         }
 
         Scope(state: \.settingsTab, action: \.settingsTab) {
-            MyPageFeature()
+            MyActivityFeature()
         }
 
         BindingReducer()
@@ -71,7 +71,7 @@ struct AppTabView: View {
                 case .home:
                     HomeView(store: store.scope(state: \.homeTab, action: \.homeTab))
                 case .settings:
-                    MyPageView(store: store.scope(state: \.settingsTab, action: \.settingsTab))
+                    MyActivityView(store: store.scope(state: \.settingsTab, action: \.settingsTab))
                 }
                 CustomTabView(selectedItem: $selectedItem)
                     .padding(.bottom, 20)
